Fix for passing tests
Now all tests pass. It should be stated somewhere in the documentation that the function repeat can accept a string as argument "num", it is not not very intuitive, and also can leads to problem if the string does not represent an integer. Now in that case a TypeError is thrown. It can also be passed as argument a negative integer and it does not make any sense actually, so now even in that case a TypeError is thrown. Added check for null and undefined cases, now needed due to the new check on argument num