[Refactor]: Repeat String Module
Overall propose that we simplify this module. First, we check and validate
incoming arguments. 'string' has to be a string and als 'count' should
always be an integer >= 0

Then just use a simple for loop to make sure that we process and give out
the result for the user.

Also, please note that String.substr is being deprecated as per MDN.
Checkout https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/substr